EditableField class moved to DataEditor project
