+++ Core/rdrun.h
@@ -68,20 +68,6 @@ struct EdExitD
 	/*"Q" quit   #0 dummy*/
 };
 
-struct EditableField
-{
-	FieldDescr* FldD = nullptr;
-	std::vector<LogicControl*> Checks;
-	FrmlElem* Impl = nullptr;
-	std::vector<Dependency*> Dependencies;
-	std::vector<XKey*> KL;
-	uint8_t Page = 0, Col = 0, Ln = 0, L = 0;
-	WORD ScanNr = 0;
-	bool Tab = false, Dupl = false, Used = false;
-	bool EdU = false, EdN = false;
-	bool Ed(bool IsNewRec);
-};
-
 struct ERecTxtD
 {
 	WORD N;
