+++ DataEditor/DataEditor.vcxproj
@@ -20,12 +20,14 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="DataEditorParams.h" />
+    <ClInclude Include="EditableField.h" />
     <ClInclude Include="EditD.h" />
     <ClInclude Include="EditReader.h" />
     <ClInclude Include="DataEditor.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="DataEditorParams.cpp" />
+    <ClCompile Include="EditableField.cpp" />
     <ClCompile Include="EditD.cpp" />
     <ClCompile Include="EditReader.cpp" />
     <ClCompile Include="DataEditor.cpp" />
