+++ DataEditor/DataEditor.vcxproj.filters
@@ -27,6 +27,9 @@
     <ClInclude Include="EditD.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="EditableField.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="EditReader.cpp">
@@ -39,6 +42,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="EditD.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="EditableField.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
