+++ DataEditor/EditableField.cpp
@@ -1,0 +1,7 @@
+#include "EditableField.h"
+#include "../fandio/FieldDescr.h"
+
+bool EditableField::Ed(bool IsNewRec)
+{
+	return FldD->isStored() && (EdU || IsNewRec && EdN);
+}
